refactor(backoffice-v2): moved end-user query to modules
