+++ apps/backoffice-v2/src/domains/individuals/queries/useEndUserByIdQuery/useEndUserByIdQuery.tsx
@@ -1,0 +1,12 @@
+import { useIsAuthenticated } from '@/domains/auth/context/AuthProvider/hooks/useIsAuthenticated/useIsAuthenticated';
+import { useQuery } from '@tanstack/react-query';
+import { endUsersQueryKeys } from '../../query-keys';
+
+export const useEndUserByIdQuery = ({ id }: { id: string }) => {
+  const isAuthenticated = useIsAuthenticated();
+
+  return useQuery({
+    ...endUsersQueryKeys.byId({ id }),
+    enabled: !!id && isAuthenticated,
+  });
+};
