+++ apps/backoffice-v2/src/domains/individuals/query-keys.ts
@@ -1,0 +1,9 @@
+import { createQueryKeys } from '@lukemorales/query-key-factory';
+import { getEndUserById } from './fetchers';
+
+export const endUsersQueryKeys = createQueryKeys('end-users', {
+  byId: ({ id }: { id: string }) => ({
+    queryKey: ['end-users', id],
+    queryFn: () => getEndUserById({ id }),
+  }),
+});
